requirements.txt
Flask
Flask-SQLAlchemy
Werkzeug
soundfile
openai-whisper
numpy
